refactor subsites in cms area to extension
